pyproject.toml
[tool.poetry]
name = "tinydenticon"
version = "0.1.0"
description = "identicon.js compatible Identicon implementation in Python 3"
authors = ["tinytengu <tinytengu@yandex.ru>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
webcolors = "^1.12"
pycryptodomex = "^3.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
